\name{bDst}
\alias{bDst}
\docType{data}
\title{
Wood basal density data for a few tree species
}
\description{
Wood basal density is used to convert wood volume measure in the field to its dried weight. It is measured in \eqn{\frac{kg}{m^3}}.
}
\usage{data(bDst)}
\format{
A data frame with 13 observations on the following variable.
\describe{
  \item{\code{basalDensity}}{a numeric vector}
  }
}
\details{
}
\source{
\url{http://www.ricercaforestale.it/index.php?module=CMpro&func=viewpage&pageid=713}
}
\examples{
data(bDst)
}
\keyword{datasets}
