\name{plotPolarSegment}
\alias{plotPolarSegment}
\title{Plots a segment}
\usage{
  plotPolarSegment(a0, d0, a1, d1)
}
\arguments{
  \item{a0}{angle of first set of coordinates}

  \item{d0}{distance of first set of coordinates}

  \item{a1}{angle of second set of coordinates}

  \item{d1}{distance of second set of coordinates}
}
\value{
  \code{NULL}
}
\description{
  Plots a segmente given two set of polar coordinates
  (angle, distance from tree base), may be used to
  represent buildings close to the tree on the CM plot
}

