\name{powerEquation}
\alias{powerEquation}
\title{Applies the exponential equation Y = a * X^b...}
\usage{powerEquation(a, b, x)
}
\description{Applies the exponential equation \eqn{Y = a * X^b}
given \eqn{a}, \eqn{b} and \eqn{X}}
\value{the dependent variable (y)}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{a}{the parameter a (a real number) in the exponential equation}
\item{b}{the parameter b (a real number) in the exponential equation}
\item{x}{the dependent variable (a real number)}
}

