\name{pureQuadraticEquation}
\alias{pureQuadraticEquation}
\title{Applies the pure quadratic equation Y = a + bX^2...}
\usage{pureQuadraticEquation(a, b, x)
}
\description{Applies the pure quadratic equation \eqn{Y = a + bX^2}
given \eqn{a}, \eqn{b} and \eqn{X}}
\value{the dependent variable (y)}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{a}{the parameter a (a real number) in the pure quadratic equation}
\item{b}{the parameter b (a real number) in the pure quadratic equation}
\item{x}{the dependent variable (a real number)}
}

