\name{setBranchesCM}
\alias{setBranchesCM}
\title{Stores branches CM...}
\usage{setBranchesCM(object, value)
}
\description{Stores branches CM}
\details{Method "calcVectors" must be invoked to take changes into effect}
\value{the updated list}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{object}{the list holding tree data}
\item{value}{the new branches CM (0.01 to 1)}
}

