\name{summary.CM}
\alias{summary.CM}
\title{Summary of Centre of Mass data...}
\usage{summary.CM(object, ...)
}
\description{Summary of Centre of Mass data}
\details{Prints in a human-readable format the polar and cartesian coordinates of tree CM}
\value{\code{NULL}}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{object}{An object of class CM}
\item{...}{Additional arguments affecting the summary produced}
}

