\name{toPolar}
\alias{toPolar}
\title{Converts cartesian (x, y) into polar (angle, distance) coordinates...}
\usage{toPolar(x, y)
}
\description{Converts cartesian (x, y) into polar (angle, distance) coordinates}
\details{Converts cartesian coordinates (x, y) into polar (angle, distance) ones, assuming (0, 0) as origin of axes}
\value{A vector holding angle in degrees (integer) and distance (real) in the same unit as x and y}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{x}{abscissa coordinate}
\item{y}{ordinate coordinate}
}

