\name{treeVectors}
\alias{treeVectors}
\title{Computes cartesian coordinates and moments of branches and logs...}
\usage{treeVectors(object)
}
\description{Computes cartesian coordinates and moments of branches and logs}
\details{A data frame is populated width branch and log masses, along with x, y cartesian coordinates and x, y, and z moments.
z coordinates and moments are calculated only if branches height from the ground (and inclination) have been recorded in the field.}
\seealso{\code{\link{getCoordinatesAndMoment}}}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{object}{an object of class vectors}
}

