% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.R
\name{centreOfMass}
\alias{centreOfMass}
\title{Computes the centre of mass of the tree}
\usage{
centreOfMass(object)
}
\arguments{
\item{object}{A data frame of class \code{vectors}}
}
\value{
A vector holding \eqn{x}, \eqn{y}, \eqn{z} coordinates of the centre of mass
}
\description{
The \eqn{x} coordinate of the centre of mass is defined as \eqn{\frac{\sum(m_ix_i)}{\sum(m_i)}} where \eqn{m_i} is the biomass of the \eqn{i^{th}} branch and \eqn{x_i} is the \eqn{x} coordinate of the \eqn{i^{th}} branch. \eqn{y} and \eqn{z} coordinates are similarly computed.
The centre of mass computation excludes branches to be pruned (ie: those whose \code{toBePruned} value is set to \code{TRUE}).
}

