\name{powerEquation}
\alias{powerEquation}
\title{Returns the result of an exponential equation}
\usage{
  powerEquation(a, b, x)
}
\arguments{
  \item{a}{the parameter \eqn{a} in the exponential
  equation}

  \item{b}{the parameter \eqn{b} in the exponential
  equation}

  \item{x}{the independent variable}
}
\value{
  the dependent variable (\eqn{Y})
}
\description{
  Returns the result of the exponential equation \eqn{Y = a
  * X^b} given \eqn{a}, \eqn{b} and \eqn{X}
}
\author{
  Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}
}

