% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{pureQuadraticEquation}
\alias{pureQuadraticEquation}
\title{Returns the result of a pure quadratic equation}
\usage{
pureQuadraticEquation(a, b, x)
}
\arguments{
\item{a}{the parameter \eqn{a} in the pure quadratic equation}

\item{b}{the parameter \eqn{a} in the pure quadratic equation}

\item{x}{the dependent variable}
}
\value{
the dependent variable (\eqn{Y})
}
\description{
Returns the result of the pure quadratic equation \eqn{Y = a + bX^2}
given \eqn{a}, \eqn{b} and \eqn{X}
}
\seealso{
Other Biomass: \code{\link{allometryABDC}},
  \code{\link{allometryAsca2011}},
  \code{\link{allometryCutini2009}},
  \code{\link{allometryPorte2002}},
  \code{\link{logBiomass}}, \code{\link{powerEquation}},
  \code{\link{treeBiomass}}, \code{\link{treeTotalBiomass}}
}

