% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setBranchesCM}
\alias{setBranchesCM}
\title{Stores branches CM in an object}
\usage{
setBranchesCM(object, value)
}
\arguments{
\item{object}{the object of class \code{treeData}}

\item{value}{the new branch CM}
}
\value{
the updated list
}
\description{
Stores branches CM in the object provided as an argument. branchesCM has to be in the range 0.01 - 1
}
\note{
Method \code{\link{treeVectors}} must be invoked to take changes into effect
}

