% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.CM}
\alias{summary.CM}
\title{Summary of Centre of Mass data}
\usage{
\method{summary}{CM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{CM}}

\item{...}{Additional arguments, not used}
}
\value{
\code{NULL}
}
\description{
Prints in a human-readable format the polar and cartesian coordinates of tree CM
}

