\name{switchBranchPruningStatus}
\alias{switchBranchPruningStatus}
\title{Sets pruning status of a branch...}
\usage{switchBranchPruningStatus(object, value)
}
\description{Sets pruning status of a branch}
\details{Switches pruning status of a record in the raw data.frame from field measures
Method "calcVectors" must be invoked to take changes into effect}
\value{the updated list}
\author{Marco Bascietto \email{marco.bascietto@ibaf.cnr.it}}
\arguments{\item{object}{the list holding tree data}
\item{value}{number of row or vector of numbers of rows to be switched}
}

